# Improve performance of scanning and serving