--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "comics"
-version = "2.2.15"
+version = "2.2.16"
 edition = "2021"
 authors = ["henry40408 <2316687+henry40408@users.noreply.github.com>"]
 description = "Simple file server for comic books"
@@ -17,6 +17,7 @@ chrono = "0.4.35"
 clap = { version = "4.5.2", features = ["derive", "env"] }
 imsz = "0.3.1"
 rand = "0.8.5"
+rayon = "1.9.0"
 rpassword = "7.3.1"
 serde = { version = "1.0.197", features = ["derive"] }
 serde_json = "1.0.114"