--- a/Cargo.toml
+++ b/Cargo.toml
@@ -16,6 +16,7 @@ blake3 = "1.5.1"
 chrono = "0.4.35"
 clap = { version = "4.5.2", features = ["derive", "env"] }
 imsz = "0.3.1"
+parking_lot = "0.12.1"
 rand = "0.8.5"
 rayon = "1.9.0"
 rpassword = "7.3.1"