# rayon

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `./target/release/comics.before list` | 135.3 ± 1.1 | 133.9 | 138.1 | 3.79 ± 0.16 |
| `./target/release/comics list` | 35.7 ± 1.4 | 32.7 | 40.7 | 1.00 |