@henry40408 2024-03-15
`cargo b --release`

before

```
Running 10s test @ http://localhost:8080
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   122.19ms    7.87ms 154.99ms   96.31%
    Req/Sec    40.69      8.82    60.00     28.00%
  814 requests in 10.01s, 70.61MB read
Requests/sec:     81.32
Transfer/sec:      7.05MB
```

after

```
Running 10s test @ http://localhost:8080
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    21.29ms    4.07ms  64.53ms   93.86%
    Req/Sec   236.43     28.46   272.00     88.00%
  4713 requests in 10.01s, 408.83MB read
Requests/sec:    470.97
Transfer/sec:     40.85MB
```