@henry40408 2024-03-16
on NAS, HDD

- before: 65675ms https://github.com/henry40408/comics/pull/65#issuecomment-1996170846
- after: 14792ms (-77.4770%)